requirement.txt
django==3.0.6
pytesseract==0.3.4
pdf2image==1.13.1
Pillow==7.0.0